Comments, Notes, Feeds, Event Templates, Event Federation, Calendar export and more
* Add public calendar, public station view, event category management, and quick field templates

- Public calendar: events/categories/fields can be marked public, iCal feed via ical4j, public event API
- Public station view: unified sidebar layout at /public/station/:uid with calendar + knowledge base tabs
- Discovery: shows all stations with public content, own station badge, public station links
- Event category management: dedicated view with reorder, max shown events, public toggle
- Quick field templates: Ort/Treffpunkt/Thema quick-add buttons in field editor
- Station settings: publicCalendarEnabled toggle
- DB: patch_5 adds public columns to station, event_category, station_event, event_field
- Demo: first station marked public with public categories
- Tests: fixed for new entity signatures

* Add fire theme and update landing page

* Reorganize templates and theming. Add color blind themes

* Update helpcenter

* Implement EventBus

* Domain event system, notification overhaul, feeds, event templates, and infra cleanup

## Domain Event System
- DomainEvent, DomainEventHandler, DomainEventBus infrastructure with Guice multibinding
- 17 event handlers registered for all notification-producing operations
- All routes fully decoupled from NotificationService (Events, News, Exchange, Forms, Procurement, MemberGroups, Lending)
- Services now publish domain events; handlers create notifications
- DemoService migrated to use services for event-producing operations

## User Feeds (iCal, RSS, Atom)
- Feed token system: user_feed_token table, FeedTokenService, FeedTokenRoutes
- User-bound iCal feed at /public/feed/{token}/events.ics
  - Filters out declined events and events with expired registration deadlines
  - 1-hour cache
- RSS 2.0 and Atom 1.0 notification feeds using ROME library
  - Per-type feed_enabled preference (new column on user_notification_settings)
  - NotificationToggle extended with 'feed' field
- Frontend: FeedSection component with token management + copyable feed URLs
- Frontend: Feed toggle column in notification settings
- Help center: RSS/Atom guide and iCal guide covering Thunderbird, Outlook, Android, iOS, web readers

## Event Templates
- event_template, event_template_field, event_template_restriction tables
- Full CRUD backend (EventTemplateRepository, EventTemplateService, EventTemplateRoutes)
- Frontend: TemplateManageView with create/edit/delete + field editor
- Frontend: "Load Template" dropdown in EventEditView that populates form from template
- Quick field templates now include isPublic flag (Ort/Treffpunkt: public, Thema: internal)
- Demo data: Standard-Übung and Wettbewerb templates

## Help Center
- Dedicated help pages for theming, sessions, notifications, modules, import
- No duplicate help components (linter check added)
- Sidebar linkage check added to linter
- RSS/Atom and iCal setup guides with platform-specific instructions

* Improve event templates, registration and notifications

* Add comments, notes, item detail view, federated events, registrations rework, and templates

## Comments & Notes System
- event_comment table with full nesting (unlimited depth)
- entity_note + entity_note_version tables with diff-based versioning
- EventCommentRepository, NoteRepository, CommentService, NoteService
- EventCommentRoutes (GET/POST/PUT/DELETE), NoteRoutes (GET/PUT/versions)
- @mention parsing (@[memberId]) with MentionedInComment domain event + notification
- Frontend: CommentThread (recursive), CommentSection, NoteEditor (with version history)
- Integrated into EventDetailView

## Item Detail View
- New route /station/inventory/item/:id
- Shows item metadata, current assignment, full assignment history, manager notes

## Federated Events (Phase 2)
- FederationHttpClient: signedPost, signedDelete, fetchFederatedEvents, register/withdraw
- FederatedEventRoutes: source station proxy (local + remote partners)
- Webhook receivers: event-registration-status, member-name-changed
- EVENT_SHARE capability added to federation system
- WebhookEvent types: EVENT_REGISTRATION_ACCEPTED, EVENT_REGISTRATION_DENIED, MEMBER_NAME_CHANGED

## Event Templates (extended)
- Replaced layouts entirely (templates are a superset)
- Added attendanceTemplateId + registrationLimit to templates
- Dedicated TemplateEditView (own page, not modal)
- Load Template dropdown in EventEditView

## Registrations Rework
- RegistrationsView: grouped by event, sorted by deadline, expandable fair acceptance table
- registrationLimit field on events (DB + backend + frontend)
- RegistrationDeadlineChecker: notifies managers when deadline expires with pending registrations
- Persistent via deadline_notified column (survives restarts)

## Layouts Removed
- All layout routes, views, sidebar links removed
- Templates fully replace layout functionality

* Add @mention autocomplete, news comment migration, user profile notes

   ## MentionInput Component
   - Textarea with @mention autocomplete dropdown
   - Triggers on @ keystroke, filters members as user types
   - Inserts @[id:name] token format
   - Integrated into CommentSection and CommentThread (replaces plain TextAreaInput)

   ## News Comment Migration
   - Created NewsCommentSection that adapts news API to generic Comment interface
   - Replaced inline comment rendering in news ListView with NewsCommentSection
   - Removed ~80 lines of inline comment logic (submitComment, handleReply, handleEdit, handleDelete, loadComments)
   - News comments now use same CommentThread with full nesting and @mention support

   ## User Profile Notes
   - NoteEditor integrated into member DetailView (visible to managers only)
   - Entity type: MEMBER, diff-based versioning

   ## NoteEditor Component
   - Markdown textarea with save button (tracks changes)
   - Version history panel with diff patches
   - Used by ItemDetailView (ITEM notes) and member DetailView (MEMBER notes)

* Dump current state on unit tests

* Finish test extension

* Refactoring

* Refactoring

* Refactor json strings to actual objects

* Update workflows

* Refactor constants

* Apply spotless

* Fix wrong constant

* Test event handler

* Update changelog

+++ src/main/java/dev/chojo/ember/feature/quiz/repository/QuizTestRepository.java
@@ -233,7 +233,7 @@ public class QuizTestRepository {
                 .first();
     }
 
-    public QuizTestAttempt createAttempt(int testId, int memberId, int maxPoints) {
+    public QuizTestAttempt createAttempt(int testId, int memberId, double maxPoints) {
         return Query.query("""
                         INSERT INTO quiz_test_attempt(test_id, member_id, max_points)
                         VALUES (:test_id, :member_id, :max_points)
